Arel::Crud is not included in Arel::Table in rails7, prepend it explicitly
